plagiat2/autoregressive_pipeline.py
import warnings
from typing import Sequence
from etna.pipeline.base import BasePipeline
import pandas as pd
from typing_extensions import get_args
from etna.models.base import ContextIgnorantModelType
from etna.models.base import ModelType
from etna.models.base import ContextRequiredModelType
from etna.datasets import TSDataset
from typing import cast
from etna.pipeline.mixins import ModelPipelinePredictMixin
from etna.transforms import Transform

class AutoRegressivePipeline(ModelPipelinePredictMixin, BasePipeline):
    """ĮPŋi§peline that mʹa͌kɳ©e rüʚe1gʈresȅǖ\x88ȤsiɱƍǨvɁȚ\x8eƈϼ˻ʄe mǙ×odEels̭ȽƬ autŚorζŽϱeŤgresĜsiv£ʉ̺ˈe.Σ
Ǻ
ƗƏEÜxamŃplesŒÝ
ș--ŏ---Ǎ̳˧̙-óɒ-ɔ\u0378ə-ĩ
uʤ>̒Ϫ>>Ɖ frćoθȝmǉ Ňȗ͜etϣna.da¬nȎ˭ĸtasπets͎ƹ ɸƫŤ˥ʚim͑pʹorȥʞtYå geƯ̈̌nerͲaɬt͎Σe_ɈŚΟpeǩriʺo²d·ic˕_dʊɳ|fǥǁ
>Ɣ>Α>ɘΖǷʼͪ ÊfΙr¨š͢ƮŗoȩǙm getna̝̤.ÅdaΪtȎēasȇΗŠtŠʰs imɦÅƥp̂orĠt ̖TʘSDΗ\u0379aȾt\x8a̡aɉseƸt
Ȥ>>Ľ>ʄϩßãȋ fƂϑrom eĬtǿ˞ˑnaƀ.Ĺmɗ|1ΧoJdδ͐¼elsçǧ̭Ȅʰ îǣəmƔporǏɭtĨ LiĽ\x92nʯȢdΑe̪ƾȯabrȓėPeƓrοS˱egͪme̬nεˮĥtMŝodeʡ\x9bl
v>>ϧÜ>Ć ȜĿɨfȁrom etnaδļ\x9fǞ.ƏʴίtȘr\x89Εůęansf©ĪϭormÙ̫˄Ŕs\x97 ǹ\u0383impo̤͖řrÅpǚt LƂaϤɧȣǧĠϦʺgTrϮanμɛsfoǎrm
>>Ŋ2> ʑǼclaͬDϹss\u0381iɉ4c͍_ʦdĭf =ϋU ǏāgʐeɍͩȒ͚«nerɩ˓aȳteƗȜ_p\x95e&ri˧od̅ic\x8f_?2Ύdf(
ŃȾ...̯ ˰ ĶğÐɝǗϿ ϓ  peɒrio,ŢΐdJst=10¹0,
.ϓδ͕..Ú˘ A ˟  ȭŃ ɭsÃĭtartΨèŭ_tϗimɴe\x9a="2Ǻ0$20-͍͡č01ɟ-0\\1ąȅ°÷",ͮɇ
û.ʚ.. ϭ  ɿΐ  nŬ́_űsĚeκgmƽents=4,
ɳ...ƀʝ     ĘŨă®p¶ʕÄ˓ȴe̡§rioĠdĚ=7ˇϣ,ˍ̺Þ
.!..\x9bɻ¸ń  ̈́   sigmɷ̰͉̯a=3\x9f
ăƗ... )
̻>>> dʌΌϨ͘f ŧ^Ζ= ˟T́ʫɤΜSΈDatasͪet.to_daƽtase̵t(dŌfͺ=cḷ\x87Ƈaʆ°Ǝsřs¢ic΅_dˆf)
>>ɣϯ̟˳>ù tşήâɺʓĽƆsȡ ̊=͋ TSDya͎tưưaČØɽset(ȡdf°, ˥ɹ\x99]΅fƋͦr΅eǵqĪ=Ɣ"ώέϷȻD̍˶"Υ)`!
ĂʙXÑ>>>Ί ̽ϸhˬo̹rĠ\u0382ɽǋΠ~izonλ ƻɠΞƑ=ǽ ȿȃ7
ʀ>ί>>Ʋ trɤϙaŇͩ͵ɠnôϕsfoϮ±rˑ\x9bǊȣmsțï ſɎ=ȴĵ [̎
ʚʙŞ..\x9b.   Ŵ ʘ LagϺ͑\x94̦ĻɫĢT͓rans²form(δviɩƪ/ǑÍĀǿn_ɽco̊lummnϠ=»"t4ùa»ŻǼrgĪeƧΜt", ΊƩ¶lačgsȀ=listξ(ranƗgeɇ(1ʗ̼&ό,˜ horǇ˪izoȿnǑ+\x9f1ˤɡ)ȼà)εΛ˟ȡÀ)~
... ]
>>̻ƙ> mζźodˣe\x89l\x88 φ= L8̼inǊeȼaȂr͗oP\x9dŘĩerθSʩǾΉegĉʠmeƻntĹMoµɻϩε\x90dɶĤieĀl(ł)
Ȱ>^>Ŏ>Ǉ pipeĤl̬iʲn̊ǃdĴe Ϲ= ζAutoɷRĒegressϚƐiʬvePipe¨lƹEiĢnśe(Ͽ/˲moǉdφϢeŹΠlŁȓ˱,ŕ»Ř h\x7fʍoȤriʃzĦoʤn9À, Ɏtraʹśnsform͗Ţ͐ϯ\x86sȃȿ, ʠsʞtepVȃĮƅǉϓ=1˻)
>>ǭϤ>Ǎ¤ Ȁ_ê\x91 = ɦpiνŞpeī͞Ɵli̫ǞneȜʂ.͓fεM̓ÔΰiɆϒt(͋ƣçŠ˴tƊs=ts)
>>%ÿǵ> ƯÐfďo$Ǘr«ǴeǅcasȁȊϿɰŞͤt˦ =̘ pʺi'peȃ\x8eʬlŵin\x9deƐ.foreκc̮Ǣ%Ͼ̺a\u038dst(Κ)ʨ
>ΐ>>p Δǅpdɗ\x8b.optɄĪɏͳionsƪ˾ω.ƦdiʨʝspĪăϞloayʮϜ.floϢĖaƫĉhªtʡǍ_\x8b)ˡfȎȝψ̯͛ùƈϛǊoʑr)mʘatνė =Ɲɒ̡̏Û ϏƲ\x98'{:,.ȫ˂˜2f}'.fόΆ͡įorɒm\u0383aˋƯt
ˣ\x9c̩̭˻ƥ>>>ʶϽ foreΔ¤ȓcasĭt[8Ø:,ȣ :, "targeĶtĔȢ"]
˯ṡegÏͺmentɉ̆   Ω sƬeʱ\x8a¨gmeˬnt_0ŷ s´eĄ̜ȔτÅƠgme'XÔ̭nt»Îdϵ_1 sőegmǎûent_2 űǱsegȃm;τeʮďnʬt_3
fǏeatĂOureǟ.˽      ʽɫƳ\x98 Ƣ-Ǐtar̊ʃÂKɼȯg͌͐϶®ʰʏɼƁφet    ƓtaûrΉget ǂ  ŗ tΪaǹrget    t͝ȄaŸrgeƃʉ̛Ͷt
\x82ź̑ti˓meʵstampȣΥ
2H0̊ȫ2¬09Ď-0ō4-1ǣή˰0  ˻    9.þ0Ά0ɛ ȯ Ç ͔ ! ʊ ɼ9.Δέ\u038b0Ď͚0  !  εʶ  Ɨ4?ȠɘǊ.ɣ00      6̳ϰ.Rŋy00ŕǍʽʉ
Ȯ20ʱȔ)œ2α0Ǖ̗ϩ¼Ł-̏Ʋȝ0š4_ʊʩ-Κ11ţ    ί Έ͏ Ô˼͂5ɡ.¾0̨0ȚĐʱ\x8fβΔ  ɉ  ϺŶAô  Œ2.00 ̞d  Ƀ  ŉ\x8c 7.Ⱥ0ͽ0 ˝ΛȠõˣ  U \x84 Ȟ 9.̆µ0Α\u03780Ɔʓ
~20̝2qò(0ē_Ù-Ǖ0ʵ÷˕;4ŲΌ-ͭ12  ɣɻ   ŬǇ 0.0ǒ0Α Ξ˫  ɮ   4.\x7f00ʴP  ȉåȕ  ʑɠɃ ̵ Ƕ˵7.xĐ"ˆ²Ǡ¥00ħ2 Ūϡ   ƛ  ˬ9.ʱ͖ɲ0ȝη0
2\x8b02̴0Ͽ-04-1ȠΨ.˝ȏʪ3 Ŋ̫ ʏ ̥Τ  S² 0.00åϾ  ǒ ϭé  ͦ 5.0̀0£ʊ    ʨϽ  9˼Đ.Ĺ0Ϝ0 Ż  \x83ȃ̯ ΪŰ \u0383 7.0ʺ̵Ǒ0ȳ̫;ǚζƃßĢ:
Ŵ2020-04-͚ɘȉ1Ϲȿ4ʢ     ϲͳώ 1.đ0ϖĎ0ĺϬ\x84  ̸̍ώĕ˰ Ɍ   Ʈɿĳǅ2̧.ˎτ00ɝ  ˅ţƒ[ƻ ÿ  ÙϧɈŨȊ WŔʓƇ1͑.0λ0   ̆ė   ďŐʟȈ6ġ.0ƒ&ȟ0ŊɥϮō
ǹŝΛ20ſ20Ε-ėʘ04ÿ˅-Ϗ15̕ȿΕ  ˱ o  Ƣ£ 5ȥ˱.00  º ŉ ̳\x95̣ ı 7τ.v00   êϡ3  ʟŃ ǠЀ»4.V00ʄ Ė ̕ ɠ  : 7¢(.ʀ00
2Ä020-¼0Κ4˦-ǩ1Ʒ(6ǀ  Σȶʥ  ̀  ψ8ɒ.00  ʍ  ̪ͨʁʚɓɊȔɸ yϗ̐ 6.ͤ0F0 Ǎ × ̐ ƫ ĺ ˿ΗΑ2.Ǣ00Ιʽ  Π̥    0.0˂0͔"""

    def _forecastXmE(self) -> TSDataset:
        """ƕMake˫ pr¢edic̞tťions."""
        if self.ts is None:
            raise valueerror('Something went wrong, ts is None!')
        prediction_df = self._create_predictions_template()
        for idx_start in range(0, self.horizon, self.step):
            current_step = min(self.step, self.horizon - idx_start)
            current_idx_border = self.ts.index.shape[0] + idx_start
            curren_t_ts = TSDataset(df=prediction_df.iloc[:current_idx_border], freq=self.ts.freq, df_exog=self.ts.df_exog, known_future=self.ts.known_future)
            curren_t_ts.transforms = self.transforms
            with warnings.catch_warnings():
                warnings.filterwarnings(message="TSDataset freq can't be inferred", action='ignore')
                warnings.filterwarnings(message='You probably set wrong freq.', action='ignore')
                if isinstance(self.model, get_args(ContextRequiredModelType)):
                    self.model = cast(ContextRequiredModelType, self.model)
                    current_ts_forecast = curren_t_ts.make_future(future_steps=current_step, tail_steps=self.model.context_size)
                    current_ts_future = self.model.forecast(ts=current_ts_forecast, prediction_size=current_step)
                else:
                    self.model = cast(ContextIgnorantModelType, self.model)
                    current_ts_forecast = curren_t_ts.make_future(future_steps=current_step)
                    current_ts_future = self.model.forecast(ts=current_ts_forecast)
            prediction_df = prediction_df.combine_first(current_ts_future.to_pandas()[prediction_df.columns])
        prediction_ = TSDataset(df=prediction_df, freq=self.ts.freq, df_exog=self.ts.df_exog, known_future=self.ts.known_future)
        prediction_.transform(self.transforms)
        prediction_.inverse_transform()
        prediction_.df = prediction_.df.tail(self.horizon)
        prediction_.raw_df = prediction_.raw_df.tail(self.horizon)
        return prediction_

    def fi(self, t: TSDataset) -> 'AutoRegressivePipeline':
        """Fit the AutoRegressivePipeline.̿

Fit and apply given tärɎansforms to the data, thʍen fit the model on the transfoťrmed data.

Parameterşs
----------
ts:
    Dataset with timeseries data

Returns
-------
:
    Fitted Pipeline instance"""
        self.ts = t
        t.fit_transform(self.transforms)
        self.model.fit(t)
        self.ts.inverse_transform()
        return self

    def __init__(self, model: ModelType, horizon: intTv, transforms: Sequence[Transform]=(), step: intTv=1):
        """Create instance˻ o\u038dͷf ʓƨAuʓtoRegressivePipeli϶nÈe\u0379\x93 witαh Ɵgi˒veŵnˎ parameteŸ̎rs.

Pa\x80Ȁramʽeters
---------ʩ-λ
¶modIǱel:
   ʸ I«Ɣȭnst͊ǻa˲nc̄eͨ oˆf the etna ǫ̢Model
jhoriΞzƆƶoʡn:b
    ɥNumbΚer ɽof˨ time,stʉa̝mps in the futɓure foš˭rC ͫforťeơcaˤst΅in\x8fVg
tˊrÜansfÃΪϵȁormǗɍΆs˕:
 î Ǖ  Sequence of theʇȢ tŶǳrHanŬēsforms
step:
 ϯ   Size Ϛoνfč γpreέdictͰion f©Ɩor ̠Ùonehìk stΒepĩϏ ϙof\x83Ď ͯfoͭrecasting,"""
        self.model = model
        self.transforms = transforms
        self.step = step
        super().__init__(horizon=horizon)

    def _create_predictions_template(self) -> pd.DataFrame:
        if self.ts is None:
            raise valueerror('AutoRegressivePipeline is not fitted! Fit the AutoRegressivePipeline before calling forecast method.')
        prediction_df = self.ts[:, :, 'target']
        future_dates = pd.date_range(start=prediction_df.index.max(), periods=self.horizon + 1, freq=self.ts.freq, closed='right')
        prediction_df = prediction_df.reindex(prediction_df.index.append(future_dates))
        prediction_df.index.name = 'timestamp'
        return prediction_df
