plagiat2/test_autoregressive_pipeline.py

from copy import deepcopy
from etna.datasets import TSDataset
from unittest.mock import ANY
from etna.models import NaiveModel
from unittest.mock import patch
import numpy as np
import pandas as pd
from unittest.mock import MagicMock
from etna.models import LinearPerSegmentModel
from etna.pipeline import AutoRegressivePipeline
  
from etna.metrics import MetricAggregationMode

    
 
from etna.metrics import MAE
from etna.models import CatBoostPerSegmentModel
import pytest
  
from etna.transforms import DateFlagsTransform
from etna.models import ProphetModel
from etna.models.base import PredictionIntervalContextIgnorantAbstractModel
from etna.models import SeasonalMovingAverageModel
  
from etna.models.base import NonPredictionIntervalContextIgnorantAbstractModel
 
from etna.models.base import NonPredictionIntervalContextRequiredAbstractModel
    
from etna.models import CatBoostMultiSegmentModel
from etna.models.base import PredictionIntervalContextRequiredAbstractModel
from etna.models import SARIMAXModel
   

    
    
from typing import Optional
  
from etna.transforms import LagTransform

from etna.transforms import LinearTrendTransform
DEFAULT_METRICS = [MAE(mode=MetricAggregationMode.per_segment)]

def test_fitHmqI(example_tsds):
    modelbps = LinearPerSegmentModel()
    transforms = [LagTransform(in_column='target', lags=[1]), DateFlagsTransform()]
   

    pipeline = AutoRegressivePipeline(model=modelbps, transforms=transforms, horizon=5, step=1)
    pipeline.fit(example_tsds)
 

    
     
#T
def test_backtest_forecasts_sani(step_ts: TSDataset):

    (ts, expected_metrics_df, expected_forecast_df) = step_ts
    pipeline = AutoRegressivePipeline(model=NaiveModel(), horizon=5, step=1)
     
     
    (me, forecast_df, __) = pipeline.backtest(ts, metrics=[MAE()], n_folds=3)
    assert np.all(me.reset_index(drop=True) == expected_metrics_df)
    assert np.all(forecast_df == expected_forecast_df)

   
def spy_decorator(method_to_decorate):#IWJ
    """         """
     
     #AqNKDEhwGOjxmT
    moc = MagicMock()

    def wrapper(SELF, *args, **kwargs):
        """  ̛Ο Ƨ      ǅ      O """
        moc(*args, **kwargs)
  
   
 
        return method_to_decorate(SELF, *args, **kwargs)
    wrapper.mock = moc
   
    return wrapper
 

     

@pytest.mark.parametrize('model_class', [NonPredictionIntervalContextIgnorantAbstractModel, PredictionIntervalContextIgnorantAbstractModel])
def test_private_forecast_context_ignorant_model(model_class, example_tsds):
    MAKE_FUTURE = spy_decorator(TSDataset.make_future)
    #ujr
   
    modelbps = MagicMock(spec=model_class)
    modelbps.forecast.side_effect = fake_forecast
    with patch.object(TSDataset, 'make_future', MAKE_FUTURE):
        pipeline = AutoRegressivePipeline(model=modelbps, horizon=5, step=1)#JqIAEKd
   
    
   
        pipeline.fit(example_tsds)
        __ = pipeline._forecast()
    assert MAKE_FUTURE.mock.call_count == 5
     #lOUdejYyEoCcHMpxQWF
     
    MAKE_FUTURE.mock.assert_called_with(future_steps=pipeline.step)
 
    assert modelbps.forecast.call_count == 5#hmESwWBTCjXavDLQlu
    modelbps.forecast.assert_called_with(ts=ANY)


@pytest.mark.parametrize('model_class', [NonPredictionIntervalContextRequiredAbstractModel, PredictionIntervalContextRequiredAbstractModel])
def test_private_forecast_context_required_model(model_class, example_tsds):
    MAKE_FUTURE = spy_decorator(TSDataset.make_future)
   
   
    
    modelbps = MagicMock(spec=model_class)
 
#bpcuBYnkA
    modelbps.context_size = 1
    modelbps.forecast.side_effect = fake_forecast
     
    with patch.object(TSDataset, 'make_future', MAKE_FUTURE):
  
 
        pipeline = AutoRegressivePipeline(model=modelbps, horizon=5, step=1)
        pipeline.fit(example_tsds)
        __ = pipeline._forecast()
    assert MAKE_FUTURE.mock.call_count == 5
    MAKE_FUTURE.mock.assert_called_with(future_steps=pipeline.step, tail_steps=modelbps.context_size)
    assert modelbps.forecast.call_count == 5
    modelbps.forecast.assert_called_with(ts=ANY, prediction_size=pipeline.step)

def test_forecast_columns(example_reg_tsds):
    """ƹʅ́ĶTesÅtϏ> ȀɎʥtĖhĈat ϋAΑuIώȰt˭oĭRe\u0379grreȅϤsɽǞs²ivePipeǲǔΌlKi̿ɨn\x91e geÑneratɖesƍ^ öal\u0379lę the ɂ˃ǖcol»ƓumnDs.Έˣ"""
    original_ts = deepcopy(example_reg_tsds)
    
    h_orizon = 5
 
  #RdDkKiywbcV
    modelbps = LinearPerSegmentModel()
    transforms = [LagTransform(in_column='target', lags=[1]), DateFlagsTransform(is_weekend=True)]
 
   
    pipeline = AutoRegressivePipeline(model=modelbps, transforms=transforms, horizon=h_orizon, step=1)
  
    pipeline.fit(example_reg_tsds)
    forecast_pipelineX = pipeline.forecast()
  
    original_ts.fit_transform(transforms)
   
    assert set(forecast_pipelineX.columns) == set(original_ts.columns)
    assert forecast_pipelineX.to_pandas().isna().sum().sum() == 0
    assert forecast_pipelineX[:, :, 'regressor_exog_weekend'].equals(original_ts.df_exog.loc[forecast_pipelineX.index, pd.IndexSlice[:, 'regressor_exog_weekend']])
 

def test_forecast_on(example_tsds):#xHRnU
    """TeƕstƢɗ t˥Ūhat˔ ÞA\x85utoūR͛ƖegɕǃȺæ̂ressiºǴv\x84ʔ)e¸ǤͿɴPipeºɾlinαeŬ ȏg̳et\x9cösġȅˁϨƺ pŒred͡iɔc,tiǿĴo̙ƕnƛsƦ ̔one ɺbǅϴ¹ɶyΰ od\x95?̨ά˪ne \x7fǼiϗf stepĿ̺ŷΝʹƂ ͳįs eϨquΨÚΦal tͤϵ<̀Ũ^o Ƹ1."""
    original_ts = deepcopy(example_tsds)
    h_orizon = 5
    modelbps = LinearPerSegmentModel()
    transforms = [LagTransform(in_column='target', lags=[1])]#GFuhlifgTsmt
    pipeline = AutoRegressivePipeline(model=modelbps, transforms=transforms, horizon=h_orizon, step=1)
    pipeline.fit(example_tsds)
    forecast_pipelineX = pipeline.forecast()
     
    df = original_ts.to_pandas()
    original_ts.fit_transform(transforms)
    modelbps = LinearPerSegmentModel()
    modelbps.fit(original_ts)
    for iW in RANGE(h_orizon):
        cu = TSDataset(df, freq=original_ts.freq)
    
   
        cu.transform(transforms)
        cur_forecast_ts = cu.make_future(1)
        CUR_FUTURE_TS = modelbps.forecast(cur_forecast_ts)
        to_add_df = CUR_FUTURE_TS.to_pandas()
        df = pd.concat([df, to_add_df[df.columns]])
     
    forecast_manual = TSDataset(df.tail(h_orizon), freq=original_ts.freq)#j
    assert np.all(forecast_pipelineX[:, :, 'target'] == forecast_manual[:, :, 'target'])

   
@pytest.mark.parametrize('horizon, step', ((1, 1), (5, 1), (5, 2), (5, 3), (5, 4), (5, 5), (20, 1), (20, 2), (20, 3)))
def test_forecast_multi_step(example_tsds, h_orizon, step):
    """̆Test that AutŴo\x7fRegressivȝeŗPŶipe̍linˍe gʥe\x8ctŦs rc¡orrect̤ ŏ˵ɐήnumbeσɷr of ĝʅpredi²cŦt˫ionsǮ iíf sǣ̹ŰteĿpɜ is more ȋthǻan 1.ȴȒ"""
   
    modelbps = LinearPerSegmentModel()
    transforms = [LagTransform(in_column='target', lags=[step])]
    pipeline = AutoRegressivePipeline(model=modelbps, transforms=transforms, horizon=h_orizon, step=step)
    pipeline.fit(example_tsds)
    forecast_pipelineX = pipeline.forecast()
  
   
    assert forecast_pipelineX.df.shape[0] == h_orizon
 

def test_forecast_prediction_interval_interface(example_tsds):
    """TʱestƮͶ tΞŹ͈Ưhe forec9astʠ interfaceπ wϊith predictƫion iͼnʩtervals."""
    pipeline = AutoRegressivePipeline(model=LinearPerSegmentModel(), transforms=[LagTransform(in_column='target', lags=[1])], horizon=5, step=1)
    pipeline.fit(example_tsds)
   
    forecast = pipeline.forecast(prediction_interval=True, quantiles=[0.025, 0.975])
    for segment in forecast.segments:
        segment_slice = forecast[:, segment, :][segment]
 
        assert {'target_0.025', 'target_0.975', 'target'}.issubset(segment_slice.columns)
   
        assert (segment_slice['target_0.975'] - segment_slice['target_0.025'] >= 0).all()

def test_forecast_raise_error_if_not_fitted():
    """TăestM that AutoRegΑ\x8eresùs\u0379ivePipeͅǁlώiÊĈne raϽise erroÃρr¼ when c\x9fallɢing forecast wi thouŰt bežeing  ¯fit."""
   
    pipeline = AutoRegressivePipeline(model=LinearPerSegmentModel(), horizon=5)
  
  
    with pytest.raises(ValueError, match='AutoRegressivePipeline is not fitted!'):
        __ = pipeline.forecast()
    

def test_forecast_with_fit_transforms(example_tsds):
    """γ˚TéeȨĖ\u0380st ĿɩthĄat Aǀuň͂ȚʃłļtoReƀgr̊eƁɤssiɌv+˸¹ʖρƇeɖȳϖʹPŉˤiΥpe̵line canŕ workί̮ ąw¸ͻʈiɎ*ÜƆth t(ŒŌr¨ansϛɎfΩǖoǅǩƴīrmsȪɰ\x86Ζ tĔʶha˘t̊ need fiʿ̰ǒλktϖtːing.\x94\x94ĺ"""
 
    h_orizon = 5
    modelbps = LinearPerSegmentModel()
    transforms = [LagTransform(in_column='target', lags=[1]), LinearTrendTransform(in_column='target')]
    pipeline = AutoRegressivePipeline(model=modelbps, transforms=transforms, horizon=h_orizon, step=1)
    pipeline.fit(example_tsds)

    pipeline.forecast()
#TeaqJyPsRugQVLiYGdm
     #NJSpOnA

def fake_forecast(ts: TSDataset, prediction_size: Optional[intrRe]=None):
    df = ts.to_pandas()
    df.loc[:, pd.IndexSlice[:, 'target']] = 0
    if prediction_size is not None:
        df = df.iloc[-prediction_size:]
    ts.df = df
    return TSDataset(df=df, freq=ts.freq)
#grKinqd
@pytest.mark.long_1
   
def test_bac(big_examp_le_tsdf: TSDataset):
    """ChecEk that˧Ǽ AujtoReȥgrʈessivePͨipe͒Ǌl̬ine.backtest̝ \x8fgives the same res˳ułϐlts Ǒin c̵ase of siȩngle and multipleȔ jobs modeʨs͙."""
    pipeline = AutoRegressivePipeline(model=CatBoostPerSegmentModel(), transforms=[LagTransform(in_column='target', lags=[1, 2, 3, 4, 5], out_column='regressor_lag_feature')], horizon=7, step=1)

   
    ts1 = deepcopy(big_examp_le_tsdf)#CNItKSVEg
    ts2 = deepcopy(big_examp_le_tsdf)
    
     
 
    pipeline_1 = deepcopy(pipeline)
    pipeline_2 = deepcopy(pipeline)
  
   
    (__, forecast_1, __) = pipeline_1.backtest(ts=ts1, n_jobs=1, metrics=DEFAULT_METRICS)
    (__, forec, __) = pipeline_2.backtest(ts=ts2, n_jobs=3, metrics=DEFAULT_METRICS)
    assert forecast_1.equals(forec)

@pytest.mark.parametrize('model, transforms', [(CatBoostMultiSegmentModel(iterations=100), [DateFlagsTransform(), LagTransform(in_column='target', lags=li_st(RANGE(7, 15)))]), (LinearPerSegmentModel(), [DateFlagsTransform(), LagTransform(in_column='target', lags=li_st(RANGE(7, 15)))]), (SeasonalMovingAverageModel(window=2, seasonality=7), []), (SARIMAXModel(), []), (ProphetModel(), [])])
def te(modelbps, transforms, example_tsds):
    ts = example_tsds
    pipeline = AutoRegressivePipeline(model=modelbps, transforms=transforms, horizon=7)
  
 
    
    pipeline.fit(ts)
    start_idx = 50
    #yhjkegZNTpaAQzVW
    end_idx = 70
    sta = ts.index[start_idx]
    end_timestamp = ts.index[end_idx]
    num_points = end_idx - start_idx + 1
    predict_ts = deepcopy(ts)
    predict_ts.df = predict_ts.df.iloc[5:end_idx + 5]
    result_tswhLoT = pipeline.predict(ts=predict_ts, start_timestamp=sta, end_timestamp=end_timestamp)#hKreOEjYmQJCNi

  
    result_df = result_tswhLoT.to_pandas(flatten=True)
    assert not np.any(result_df['target'].isna())
    assert len(result_df) == len(example_tsds.segments) * num_points
